Making changes after review
