+++ tests/Data/ProgrammesDb/EntityRepository/SegmentEventRepository/FindBySegmentTest.php
@@ -45,9 +45,9 @@ class FindBySegmentTest extends AbstractDatabaseTest
     public function findBySegmentData()
     {
         return [
-            [['s0000001'], 50, 0, ['sv000001', 'sv000003', 'sv000004', 'sv000005']],
+            [['s0000001'], 50, 0, ['sv000001', 'sv000003', 'sv000004', 'sv000005']], // Implicitly testing enabled embargoed filter
             [['s0000001'], 2, 1, ['sv000003', 'sv000004']],
-            [['s0000002'], 50, 0, ['sv000007', 'sv000008']] // Test Distinct Version
+            [['s0000002'], 50, 0, ['sv000007', 'sv000008']], // Test Distinct Version
         ];
     }
 
@@ -78,7 +78,36 @@ class FindBySegmentTest extends AbstractDatabaseTest
         // Expect embargoed version to be last (ORDER BY hasBroadcast ASC)
         $this->assertEquals($expectedPids, array_column($entities, 'pid'));
 
-        // findByPidFull query only
+        // findByPid query only
         $this->assertCount(1, $this->getDbQueries());
+    }
+
+    public function testFindBySegmentOrderByBroadcast()
+    {
+        $this->loadFixtures(['SegmentEventsBroadcastsOrderFixture']);
+        $repo = $this->getEntityManager()->getRepository('ProgrammesPagesService:SegmentEvent');
+
+        // Specific case where we load up a different fixture so we have to define the data here
+        $findBySegmentData = [
+            [['s0000001'], 50, 0, ['sv000004', 'sv000005', 'sv000001', 'sv000006', 'sv000003']],
+        ]; // Test Broadcast Order
+
+        foreach ($findBySegmentData as $data) {
+            list($pids, $limit, $offset, $expectedPids) = $data;
+
+            $ids = array_map(function ($dbId) {
+                return $this->getDbIdFromPid($dbId, 'Segment');
+            }, $pids);
+
+            $entities = $repo->findBySegment($ids, $limit, $offset);
+
+            // Expect embargoed version to be last (ORDER BY hasBroadcast ASC)
+            $this->assertEquals($expectedPids, array_column($entities, 'pid'));
+
+            // findBySegment query only
+            $this->assertCount(1, $this->getDbQueries());
+
+            $this->resetDbQueryLogger();
+        }
     }
 }
