+++ tests/Service/SegmentsService/FindByPidTest.php
@@ -4,32 +4,32 @@ namespace Tests\BBC\ProgrammesPagesService\Service\SegmentsService;
 
 use BBC\ProgrammesPagesService\Domain\ValueObject\Pid;
 
-class FindByPidFullTest extends AbstractSegmentsServiceTest
+class FindByPidTest extends AbstractSegmentsServiceTest
 {
-    public function testFindByPidFull()
+    public function testFindByPid()
     {
         $pid = new Pid('s1234567');
         $dbData = ['pid' => 's1234567'];
 
         $this->mockRepository->expects($this->once())
-            ->method('findByPidFull')
+            ->method('findByPid')
             ->with($pid)
             ->willReturn($dbData);
 
-        $result = $this->service()->findByPidFull($pid);
+        $result = $this->service()->findByPid($pid);
         $this->assertEquals($this->segmentFromDbData($dbData), $result);
     }
 
-    public function testFindByPidFullEmptyData()
+    public function testFindByPidEmptyData()
     {
         $pid = new Pid('s1234567');
 
         $this->mockRepository->expects($this->once())
-            ->method('findByPidFull')
+            ->method('findByPid')
             ->with($pid)
             ->willReturn(null);
 
-        $result = $this->service()->findByPidFull($pid);
+        $result = $this->service()->findByPid($pid);
         $this->assertEquals(null, $result);
     }
 }
